New document / doctype cleanup
