--- web/app/controllers/authenticated/new/doc.js
@@ -1,11 +1,0 @@
-import Controller from "@ember/controller";
-import { inject as service } from "@ember/service";
-import { action } from "@ember/object";
-
-export default class AuthenticatedNewDocController extends Controller {
-  @service router;
-
-  queryParams = ["docType"];
-
-
-}
