+++ web/app/templates/authenticated/new.hbs
@@ -1,4 +1,3 @@
-{{! @glint-nocheck: not typesafe yet }}
 <Header />
 <section>
   <div class="x-container">
