+++ web/app/templates/authenticated/new/doc.hbs
@@ -1,3 +1,3 @@
-{{page-title (concat "Create Your " @model.docType)}}
+{{page-title (concat "Create Your " @model.name)}}
 
-<New::DocForm @docType={{@model.docType}} />
+<New::DocForm @docType={{@model.name}} />
